# docs/swagger.yaml
basePath: /v1
definitions:
  common.AppError:
    properties:
      key:
        type: string
      log:
        type: string
      message:
        type: string
      status_code:
        type: integer
    type: object
...
  common.Response:
    properties:
      data: {}
    type: object
...
  model.Blog:
    properties:
      author_id:
        type: integer
      content:
        description: HTML content
        type: string
      created_at:
        type: string
      id:
        type: integer
      image:
        $ref: '#/definitions/common.Image'
      status:
        description: draft, published
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  model.BlogWithAuthor:
    properties:
      author:
        properties:
          full_name:
            type: string
          id:
            type: integer
        type: object
      author_id:
        type: integer
      content:
        description: HTML content
        type: string
      created_at:
        type: string
      id:
        type: integer
      image:
        $ref: '#/definitions/common.Image'
      status:
        description: draft, published
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  model.CategoryUpdate:
    properties:
      category_name:
        type: string
      description:
        type: string
      status:
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API documentation for Mocau Backend
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Mocau Backend API
  version: "1.0"
paths:
  /blogs:
    get:
      consumes:
      - application/json
      description: Get list of all blogs
      parameters:
      - description: Filter by status (draft, published)
        in: query
        name: status
        type: string
      - description: Filter by author ID
        in: query
        name: author_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Blog'
                  type: array
              type: object
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/common.Response'
      summary: List all blogs
      tags:
      - blogs
    post:
      consumes:
      - multipart/form-data
      description: Create a new blog with title, content, image and status. Author
        will be set automatically from logged in user.
      parameters:
      - description: Blog title
        in: formData
        name: title
        required: true
        type: string
      - description: Blog HTML content
        in: formData
        name: content
        required: true
        type: string
      - description: Blog status (draft or published)
        in: formData
        name: status
        type: string
      - description: Blog image
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Blog created successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: Create a new blog
      tags:
      - blogs
  /blogs/{id}:
    delete:
      description: Set deleted_at timestamp instead of removing the record. Only blog
        author can delete their own blog.
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.AppError'
        "403":
          description: Forbidden - not blog author
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.AppError'
      security:
      - BearerAuth: []
      summary: Soft-delete a blog by id
      tags:
      - blogs
    get:
      consumes:
      - application/json
      description: Get blog details by ID with author information
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.BlogWithAuthor'
              type: object
        "404":
          description: Blog not found
          schema:
            $ref: '#/definitions/common.Response'
      summary: Get blog by ID
      tags:
      - blogs
    put:
      consumes:
      - multipart/form-data
      description: Update blog fields by ID including image. Only blog author can
        update their own blog.
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: integer
      - description: Blog title
        in: formData
        name: title
        type: string
      - description: Blog HTML content
        in: formData
        name: content
        type: string
      - description: Blog status (draft or published)
        in: formData
        name: status
        type: string
      - description: Blog image
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  type: boolean
              type: object
        "403":
          description: Forbidden - not blog author
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Blog not found
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: Update a blog by ID
      tags:
      - blogs
  /categories:
    get:
      description: Returns all active categories by default.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/common.AppError'
      summary: List all categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create a new category with unique name. Field `status` defaults
        to `active` if omitted.
      parameters:
      - description: Category data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.CategoryCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Invalid payload or name existed
          schema:
            $ref: '#/definitions/common.AppError'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/common.AppError'
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - categories
  /categories/{id}:
    delete:
      description: Set status to `deactive` instead of removing the record.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.AppError'
      security:
      - BearerAuth: []
      summary: Soft-delete a category by id
      tags:
      - categories
    get:
      description: Return a single active category by numeric id. Soft-deleted categories
        are excluded.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/common.AppError'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/common.AppError'
      summary: Get category by id
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Partial update for category fields. Name must remain unique.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category update payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.CategoryUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Invalid payload
          schema:
            $ref: '#/definitions/common.AppError'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/common.AppError'
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - categories
  /login:
    post:
      consumes:
      - application/json
      description: Login with username/email and password to get access token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/model.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data: {}
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/common.Response'
      summary: User login
      tags:
      - users
  /order-items:
    post:
      consumes:
      - application/json
      description: Add a new item to an existing order (only for pending orders)
      parameters:
      - description: Order item information
        in: body
        name: orderItem
        required: true
        schema:
          $ref: '#/definitions/mocau-backend_module_orderitem_model.OrderItemCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.OrderItem'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: Add item to order
      tags:
      - order-items
  /order-items/{id}:
    delete:
      consumes:
      - application/json
      description: Delete order item from order (only for pending orders)
      parameters:
      - description: Order Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: Delete order item
      tags:
      - order-items
    get:
      consumes:
      - application/json
      description: Get order item details by ID
      parameters:
      - description: Order Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.OrderItem'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: Get order item by ID
      tags:
      - order-items
    put:
      consumes:
      - application/json
      description: Update order item information (only for pending orders)
      parameters:
      - description: Order Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Order item update data
        in: body
        name: orderItem
        required: true
        schema:
          $ref: '#/definitions/model.OrderItemUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.OrderItem'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: Update order item
      tags:
      - order-items
  /order-items/{id}/quantity:
    put:
      consumes:
      - application/json
      description: Update order item quantity (only for pending orders)
      parameters:
      - description: Order Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Quantity update data
        in: body
        name: quantity
        required: true
        schema:
          additionalProperties:
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.OrderItem'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: Update order item quantity
      tags:
      - order-items
  /order-items/bulk:
    delete:
      consumes:
      - application/json
      description: Delete multiple order items at once
      parameters:
      - description: Order item IDs to delete
        in: body
        name: ids
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: Bulk delete order items
      tags:
      - order-items
    post:
      consumes:
      - application/json
      description: Add multiple items to an order at once
      parameters:
      - description: Order items to create
        in: body
        name: items
        required: true
        schema:
          items:
            $ref: '#/definitions/mocau-backend_module_orderitem_model.OrderItemCreate'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: Bulk create order items
      tags:
      - order-items
    put:
      consumes:
      - application/json
      description: Update multiple order items at once
      parameters:
      - description: Order item updates
        in: body
        name: updates
        required: true
        schema:
          additionalProperties:
            $ref: '#/definitions/model.OrderItemUpdate'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: Bulk update order items
      tags:
      - order-items
  /order-items/order/{order_id}:
    get:
      consumes:
      - application/json
      description: Get all items in a specific order
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.OrderItem'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: List order items by order ID
      tags:
      - order-items
  /orders:
    get:
      consumes:
      - application/json
      description: Get list of orders with filtering
      parameters:
      - description: Order status (pending, confirmed, completed, cancelled)
        in: query
        name: status
        type: string
      - description: Processed by user ID
        in: query
        name: processed_by
        type: integer
      - description: Customer phone number
        in: query
        name: customer_phone
        type: string
      - description: Order number
        in: query
        name: order_number
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Order'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: List orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create a new order with customer information and optional order
        items
      parameters:
      - description: Order information with optional order items
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/model.OrderCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: Create a new order
      tags:
      - orders
  /orders/{id}:
    delete:
      consumes:
      - application/json
      description: Delete order (only pending orders can be deleted)
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: Delete order
      tags:
      - orders
    get:
      consumes:
      - application/json
      description: Get order details by order ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: Get order by ID
      tags:
      - orders
    put:
      consumes:
      - application/json
      description: Update order information
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Order update data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/model.OrderUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: Update order
      tags:
      - orders
  /orders/{id}/status:
    put:
      consumes:
      - application/json
      description: Update order status (pending -> confirmed -> completed)
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status update data
        in: body
        name: status
        required: true
        schema:
          properties:
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: Update order status
      tags:
      - orders
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"